requirements.txt
ag2==0.11.1
aiohttp==3.8.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.1.0
async-timeout==4.0.3
asyncer==0.0.8
attrs==25.4.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
babel==2.17.0
backoff==2.2.1
backtrader==1.9.78.123
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
bs4==0.0.2
build==1.3.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.115.2
fastjsonschema==2.21.2
filelock==3.20.3
filetype==1.2.0
finnhub-python==2.4.27
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozendict>=2.4.0
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langsmith==0.1.147
lark==1.3.1
lxml==6.0.2
Markdown==3.10.1
markdown-it-py==4.0.0
markdownify==1.2.2

# ✅ FIXED CONFLICTS
onnxruntime==1.19.2
pillow>=11.0.0
pydantic==2.10.4

MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
mplfinance==0.12.10b0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.8.1
notebook==7.5.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
openai==2.15.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.5
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20260107
pdftext==0.3.20
peewee==3.19.0
platformdirs==4.5.1
posthog==7.6.0
praw==7.8.1
prawcore==2.4.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.3
psutil==7.2.1
pure_eval==0.2.3
pyautogen==0.9.0
pycparser==3.0
pydantic-settings==2.8.1
Pygments==2.19.2
pypandoc==1.16.2
pyparsing==3.3.2
pypdf==6.6.2
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-magic==0.4.27
python-pptx==1.0.2
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
ratelimit==2.2.1
referencing==0.37.0
regex==2024.11.6
reportlab==4.4.9
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.1
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.5.0
scipy==1.17.0
seaborn==0.13.2
sec-api==1.0.32
Send2Trash==2.1.0
sentence-transformers==4.1.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.37.2

sympy>=1.13.3
ta==0.11.0
tabulate==0.9.0
tenacity==8.3.0
termcolor==3.3.0
terminado==0.18.1

threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.57.6
tushare==1.4.24
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.8.1
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.4.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
wrapt==1.17.3
xgboost==3.2.0
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
yfinance==1.1.0
zipp==3.23.0